PUBDEV-7612 after CR fixes
