+++ h2o-core/src/main/java/water/util/PojoUtils.java
@@ -719,5 +719,22 @@ public class PojoUtils {
   public static boolean equals(Object a, Object b) {
     return (a == b) || (a != null && b != null && b.equals(a));
   }
+  
+  public static String toJavaDoubleArray(double[] array) {
+    if (array == null) {
+      return "null";
+    }
+
+    SB sb = new SB();
+    sb.p("new double[] {");
+    for (int i = 0; i < array.length; i++) {
+      sb.p(" ");
+      sb.p(array[i]);
+      if (i < array.length - 1)
+        sb.p(",");
+    }
+    sb.p("}");
+    return sb.getContent();
+  }
 
 }
